chore: move example to another folder
