+++ example/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "header-example",
+  "homepage": "https://debtcollective.github.io/header",
+  "version": "0.0.0",
+  "license": "MIT",
+  "private": true,
+  "resolutions": {
+    "babel-jest": "24.7.1"
+  },
+  "dependencies": {
+    "@debtcollective/header": "link:..",
+    "@material-ui/core": "^3.9.0",
+    "@material-ui/icons": "^3.0.0",
+    "lodash": "^4.17.0",
+    "moment": "^2.24.0",
+    "prop-types": "^15.7.2",
+    "react": "16.8.6",
+    "react-dom": "16.8.6",
+    "react-popper": "^1.3.3",
+    "react-scripts": "3.0.0",
+    "styled-components": "^4.2.0"
+  },
+  "devDependencies": {
+    "cross-env": "^5.1.4"
+  },
+  "scripts": {
+    "start": "cross-env PORT=8080 react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
+}
