+++ example/public/manifest.json
@@ -1,0 +1,8 @@
+{
+  "short_name": "header-lib",
+  "name": "header-lib",
+  "start_url": "./index.html",
+  "display": "standalone",
+  "theme_color": "#000000",
+  "background_color": "#ffffff"
+}
