+++ src/appbar/AppBar.js
@@ -1,7 +1,7 @@
 // @flow
 import * as React from "react";
 import styled from "styled-components";
-import { AppBar as MUAppBar, Toolbar as MUToolbar } from "@material-ui/core/";
+import { AppBar as MUAppBar, Toolbar as MUToolbar } from "@material-ui/core";
 
 const Toolbar = styled(MUToolbar)({
   flexDirection: "row",
