+++ src/common/helpers.js
