+++ src/index.js
@@ -1,19 +1,34 @@
 // @flow
 
-import "./index.css";
-import Header from "./lib";
+import AppBar from "./appbar";
+import Menu from "./menu";
+import Notifications from "./notifications";
 import React from "react";
-import ReactDOM from "react-dom";
+import Session from "./session";
+import {
+  AlertsDropdown,
+  InboxDropdown,
+  ProfileItems,
+  UserAvatarDropdown,
+} from "./profile";
 
-const root = document.getElementById("root");
+const DebtcollectiveHeader = ({ links }: { links: Array<Link> }) => (
+  <AppBar>
+    <Menu links={links} />
+    <Session>
+      {({ user }) => (
+        <Notifications user={user}>
+          {({ alerts, messages, service }) => (
+            <ProfileItems>
+              <AlertsDropdown service={service} alerts={alerts} />
+              <InboxDropdown service={service} messages={messages} />
+              <UserAvatarDropdown user={user} />
+            </ProfileItems>
+          )}
+        </Notifications>
+      )}
+    </Session>
+  </AppBar>
+);
 
-const links = [
-  { href: "/admin/disputes", roles: ["admin"], text: "Admin" },
-  { href: "/disputes/my", text: "My Disputes" },
-  { href: "/", text: "Dispute Your Debt" },
-  { href: "https://community.debtcollective.org", text: "Community" },
-];
-
-if (root !== null) {
-  ReactDOM.render(<Header links={links} />, root);
-}
+export default DebtcollectiveHeader;
