+++ src/locales/index.js
