+++ src/menu/Menu.mdx
@@ -23,13 +23,13 @@ All related to the menu, such as: icon, toggle logic, drawer and navigation, thi
       { text: "Events", href: "https://community.debtcollective.org/calendar" },
       {
         text: "The Power Report",
-        href: "https://powerreport.debtcollective.org",
+        href: "https://powerreport.debtcollective.org"
       },
       { text: "Donate", href: "http://tools.debtcollective.org/donate" },
       {
         text: "About the Debt Collective",
-        href: "https://debtcollective.org#about",
-      },
+        href: "https://debtcollective.org#about"
+      }
     ]}
   />
 </Playground>
