+++ src/notifications/__tests__/NotificationsPanel.spec.js
