+++ src/notifications/__tests__/normaliser.spec.js
