+++ src/profile/__tests__/UserAvatarDropdown.spec.js
