+++ src/session/__tests__/Session.spec.js
