+++ src/session/__tests__/SessionService.spec.js
